requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Limiter==3.9.2
pytest==8.3.4
pytest-cov==6.0.0
python-dotenv==1.0.1
redis==5.2.1
gunicorn==23.0.0
celery==5.4.0
structlog==24.4.0